refactor: isolate hosted-local Cloudflare source snapshots
