+++ packages/hosted-local-harness/README.md
@@ -44,6 +44,11 @@ setting the internal `MURPH_DEV_SKIP_RUNNER_BUNDLE=1` flag supply both an existi
 runner bundle and prepared workspace artifacts; the public `--no-bundle` E2E
 command owns the workspace preparation described above.
 
+`src/dev-hosted-local/cloudflare-source-snapshot.ts` owns the temporary Worker
+source copy and built workspace dependency materialization. `stack.ts` prepares
+the runner bundle first, then supplies the resulting snapshot paths to Wrangler;
+startup, readiness, and teardown remain with the stack.
+
 `doctor` reports Docker daemon and Buildx prerequisites separately. Isolated
 Docker configuration selects the first candidate plugin directory containing an
 executable `docker-buildx`, so an empty or unusable earlier directory cannot hide
